Create JSONTextCompare
This file deals with the secondary text validation capacity of the utility. 
There are two ways to do it - one way utilizes a custom text comparison class, whereas the commented out method counts all the characters in the string to ensure that the highlighter highlights the correct location, as it itself cannot deal with lines and needs starting and ending character positions; thus that approach counts characters.